Address review: module format, gossip context, fork-safe validator check
