+++ beacon_node/beacon_chain/src/execution_proof_verification/mod.rs
@@ -1,0 +1,73 @@
+//! Gossip verification for the EIP-8025 `execution_proof` topic.
+
+use crate::BeaconChainError;
+use proof_engine::ProofEngineError;
+use types::{Hash256, Slot};
+
+pub mod gossip_verified_execution_proof;
+pub mod observed_execution_proofs;
+
+pub use gossip_verified_execution_proof::{
+    GossipVerificationContext, GossipVerifiedExecutionProof,
+};
+pub use observed_execution_proofs::ObservedExecutionProofs;
+
+use observed_execution_proofs::Error as ObservationError;
+
+#[derive(Debug)]
+pub enum Error {
+    /// The proof has already been seen (IGNORE).
+    ProofAlreadySeen,
+    /// A valid proof for this `(block_root, proof_type)` is already known (IGNORE).
+    ValidProofAlreadyKnown,
+    /// This validator already submitted a proof for this `(block_root, proof_type)` (IGNORE).
+    DuplicateFromValidator {
+        validator_index: u64,
+    },
+    /// The referenced beacon block is not known to fork choice (IGNORE).
+    UnknownBlockRoot {
+        beacon_block_root: Hash256,
+    },
+    /// The referenced beacon block is already finalized (IGNORE).
+    PastFinalizedSlot {
+        slot: Slot,
+        finalized_slot: Slot,
+    },
+    /// `proof_data` is empty (REJECT).
+    EmptyProofData,
+    /// The validator index does not exist (REJECT).
+    UnknownValidatorIndex(u64),
+    /// The validator is not active in the referenced block's state (REJECT).
+    ValidatorNotActive {
+        validator_index: u64,
+    },
+    /// The signature is invalid (REJECT).
+    InvalidSignature,
+    /// The proof engine rejected the proof (REJECT).
+    InvalidProof,
+    /// No proof engine is configured; the node should not be subscribed to the topic.
+    ProofEngineMissing,
+    /// The proof engine could not be reached or answered malformed (IGNORE).
+    ProofEngine(ProofEngineError),
+    BeaconChainError(Box<BeaconChainError>),
+}
+
+impl From<BeaconChainError> for Error {
+    fn from(e: BeaconChainError) -> Self {
+        Error::BeaconChainError(Box::new(e))
+    }
+}
+
+impl From<ObservationError> for Error {
+    fn from(e: ObservationError) -> Self {
+        match e {
+            ObservationError::FinalizedProof {
+                slot,
+                finalized_slot,
+            } => Error::PastFinalizedSlot {
+                slot,
+                finalized_slot,
+            },
+        }
+    }
+}
