+++ beacon_node/beacon_chain/src/execution_proof_verification/observed_execution_proofs.rs
