+++ beacon_node/beacon_chain/src/lib.rs
@@ -42,7 +42,6 @@ pub mod observed_aggregates;
 mod observed_attesters;
 pub mod observed_block_producers;
 pub mod observed_data_sidecars;
-pub mod observed_execution_proofs;
 pub mod observed_operations;
 mod observed_slashable;
 pub mod partial_data_column_assembler;
